requirements.txt
luno-python
pandas
python-dotenv
loguru
ccxt
pycoingecko
psycopg2
sqlalchemy
mysqlclient
schedule
python-binance
retry
TAcharts